+++ src/Tests/CaptainHook.Tests/Director/ReaderServiceManagement/ReaderServicesManagerTests.cs
@@ -1,18 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using CaptainHook.Application.Infrastructure.DirectorService.Remoting;
-using CaptainHook.Common;
 using CaptainHook.DirectorService.Events;
-using CaptainHook.DirectorService.Infrastructure;
 using CaptainHook.DirectorService.Infrastructure.Interfaces;
 using CaptainHook.DirectorService.ReaderServiceManagement;
 using CaptainHook.Tests.Builders;
 using Eshopworld.Core;
 using Eshopworld.Tests.Core;
-using FluentAssertions;
 using FluentAssertions.Execution;
 using Moq;
 using Xunit;
@@ -52,11 +46,11 @@ namespace CaptainHook.Tests.Director.ReaderServiceManagement
             var expectedCreatedServices = newReaders.Select(r => r.ServiceNameWithSuffix).ToArray();
             using (new AssertionScope())
             {
-                VerifyFabricClientCreateCalls(expectedCreatedServices);
-                VerifyFabricClientDeleteCalls();
+                _fabricClientMock.VerifyFabricClientCreateCalls(expectedCreatedServices);
+                _fabricClientMock.VerifyFabricClientDeleteCalls();
 
-                VerifyServiceCreatedEventPublished(expectedCreatedServices);
-                VerifyServiceDeletedEventPublished();
+                _bigBrotherMock.VerifyServiceCreatedEventPublished(expectedCreatedServices);
+                _bigBrotherMock.VerifyServiceDeletedEventPublished();
 
                 _bigBrotherMock.Verify(b => b.Publish(
                    It.Is<RefreshSubscribersEvent>(m => m.AddedCount == 3 && m.RemovedCount == 0 && m.ChangedCount == 0),
@@ -85,11 +79,11 @@ namespace CaptainHook.Tests.Director.ReaderServiceManagement
             // Assert
             using (new AssertionScope())
             {
-                VerifyFabricClientCreateCalls();
-                VerifyFabricClientDeleteCalls("service-1", "service-2", "service-3");
+                _fabricClientMock.VerifyFabricClientCreateCalls();
+                _fabricClientMock.VerifyFabricClientDeleteCalls("service-1", "service-2", "service-3");
 
-                VerifyServiceCreatedEventPublished();
-                VerifyServiceDeletedEventPublished("service-1", "service-2", "service-3");
+                _bigBrotherMock.VerifyServiceCreatedEventPublished();
+                _bigBrotherMock.VerifyServiceDeletedEventPublished("service-1", "service-2", "service-3");
 
                 _bigBrotherMock.Verify(b => b.Publish(
                    It.Is<RefreshSubscribersEvent>(m => m.AddedCount == 0 && m.RemovedCount == 3 && m.ChangedCount == 0),
@@ -122,11 +116,11 @@ namespace CaptainHook.Tests.Director.ReaderServiceManagement
 
             using (new AssertionScope())
             {
-                VerifyFabricClientCreateCalls(expectedCreatedServices);
-                VerifyFabricClientDeleteCalls(expectedDeletedServices);
+                _fabricClientMock.VerifyFabricClientCreateCalls(expectedCreatedServices);
+                _fabricClientMock.VerifyFabricClientDeleteCalls(expectedDeletedServices);
 
-                VerifyServiceCreatedEventPublished(expectedCreatedServices);
-                VerifyServiceDeletedEventPublished(expectedDeletedServices);
+                _bigBrotherMock.VerifyServiceCreatedEventPublished(expectedCreatedServices);
+                _bigBrotherMock.VerifyServiceDeletedEventPublished(expectedDeletedServices);
 
                 _bigBrotherMock.Verify(b => b.Publish(
                    It.Is<RefreshSubscribersEvent>(m => m.AddedCount == 0 && m.RemovedCount == 0 && m.ChangedCount == 3),
@@ -167,211 +161,15 @@ namespace CaptainHook.Tests.Director.ReaderServiceManagement
 
             using (new AssertionScope())
             {
-                VerifyFabricClientCreateCalls(expectedCreatedServices);
-                VerifyFabricClientDeleteCalls(expectedDeletedServices);
+                _fabricClientMock.VerifyFabricClientCreateCalls(expectedCreatedServices);
+                _fabricClientMock.VerifyFabricClientDeleteCalls(expectedDeletedServices);
 
-                VerifyServiceCreatedEventPublished(expectedCreatedServices);
-                VerifyServiceDeletedEventPublished(expectedDeletedServices);
+                _bigBrotherMock.VerifyServiceCreatedEventPublished(expectedCreatedServices);
+                _bigBrotherMock.VerifyServiceDeletedEventPublished(expectedDeletedServices);
 
                 _bigBrotherMock.Verify(b => b.Publish(
                    It.Is<RefreshSubscribersEvent>(m => m.AddedCount == 3 && m.RemovedCount == 3 && m.ChangedCount == 0),
                    It.IsAny<string>(), It.IsAny<string>(), It.IsAny<int>()), Times.Once);
-            }
-        }
-
-
-        [Fact, IsUnit]
-        public async Task RefreshReaderAsync_When_ReaderDoesNotExist_Then_ReaderIsCreated()
-        {
-            var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync())
-                .ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "other-reader") });
-
-            var readerServiceManager = CreateReaderServiceManager();
-            var result = await readerServiceManager.RefreshReaderAsync(subscriberConfig, CancellationToken.None);
-
-            using (new AssertionScope())
-            {
-                var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-                result.Should().Be(ReaderRefreshResult.Created);
-                VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                VerifyFabricClientDeleteCalls();
-                VerifyServiceCreatedEventPublished(desiredReader.ServiceNameWithSuffix);
-                VerifyServiceDeletedEventPublished();
-            }
-        }
-
-        [Fact, IsUnit]
-        public async Task RefreshReaderAsync_When_ReaderAlreadyExistsForThatConfiguration_Then_ReaderIsNotCreated()
-        {
-            var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync())
-                .ReturnsAsync(new List<string> { desiredReader.ServiceNameWithSuffix });
-
-            var readerServiceManager = CreateReaderServiceManager();
-            var result = await readerServiceManager.RefreshReaderAsync(subscriberConfig, CancellationToken.None);
-
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderRefreshResult.ReaderAlreadyExists);
-                VerifyFabricClientCreateCalls();
-                VerifyFabricClientDeleteCalls();
-                VerifyServiceCreatedEventPublished();
-                VerifyServiceDeletedEventPublished();
-            }
-        }
-
-        [Fact, IsUnit]
-        public async Task RefreshReaderAsync_When_ReaderExistsInOlderVersion_Then_ReaderIsUpdated()
-        {
-            var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            var oldReaderName = ServiceNaming.EventReaderServiceFullUri("testevent", "reader");
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync()).ReturnsAsync(new List<string> { oldReaderName });
-
-            var readerServiceManager = CreateReaderServiceManager();
-            var result = await readerServiceManager.RefreshReaderAsync(subscriberConfig, CancellationToken.None);
-
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderRefreshResult.Updated);
-                VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                VerifyFabricClientDeleteCalls(oldReaderName);
-                VerifyServiceCreatedEventPublished(desiredReader.ServiceNameWithSuffix);
-                VerifyServiceDeletedEventPublished(oldReaderName);
-            }
-        }
-
-        [Fact, IsUnit]
-        public async Task RefreshReaderAsync_When_CreateReaderFailsForNewSubscriber_Then_ResultHasFailure()
-        {
-            _fabricClientMock
-                .Setup(c => c.CreateServiceAsync(It.IsAny<ServiceCreationDescription>(), It.IsAny<CancellationToken>()))
-                .ThrowsAsync(new Exception());
-            var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            _fabricClientMock
-                .Setup(x => x.GetServiceUriListAsync())
-                .ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "other-reader") });
-
-            var readerServiceManager = CreateReaderServiceManager();
-            var result = await readerServiceManager.RefreshReaderAsync(subscriberConfig, CancellationToken.None);
-
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderRefreshResult.CreateFailed);
-                VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                VerifyFabricClientDeleteCalls();
-                VerifyServiceCreatedEventPublished();
-                VerifyServiceDeletedEventPublished();
-            }
-        }
-
-        [Fact, IsUnit]
-        public async Task RefreshReaderAsync_When_CreateReaderFailsForUpdatingSubscriber_Then_ResultHasFailure()
-        {
-            _fabricClientMock
-                .Setup(c => c.CreateServiceAsync(It.IsAny<ServiceCreationDescription>(), It.IsAny<CancellationToken>()))
-                .ThrowsAsync(new Exception());
-            var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            var oldReaderName = ServiceNaming.EventReaderServiceFullUri("testevent", "reader");
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync()).ReturnsAsync(new List<string> { oldReaderName });
-
-            var readerServiceManager = CreateReaderServiceManager();
-            var result = await readerServiceManager.RefreshReaderAsync(subscriberConfig, CancellationToken.None);
-
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderRefreshResult.CreateFailed | ReaderRefreshResult.ReaderAlreadyExists);
-                VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                VerifyFabricClientDeleteCalls();
-                VerifyServiceCreatedEventPublished();
-                VerifyServiceDeletedEventPublished();
-            }
-        }
-
-        [Fact, IsUnit]
-        public async Task RefreshReaderAsync_When_DeleteReaderFailsForUpdatingSubscriber_Then_ResultHasFailure()
-        {
-             _fabricClientMock
-                .Setup(c => c.DeleteServiceAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-                .ThrowsAsync(new Exception());
-            var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            var oldReaderName = ServiceNaming.EventReaderServiceFullUri("testevent", "reader");
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync()).ReturnsAsync(new List<string> { oldReaderName });
-
-            var readerServiceManager = CreateReaderServiceManager();
-            var result = await readerServiceManager.RefreshReaderAsync(subscriberConfig, CancellationToken.None);
-
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderRefreshResult.CreateFailed | ReaderRefreshResult.Created | ReaderRefreshResult.ReaderAlreadyExists);
-                VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                VerifyFabricClientDeleteCalls(oldReaderName);
-                VerifyServiceCreatedEventPublished(desiredReader.ServiceNameWithSuffix);
-                VerifyServiceDeletedEventPublished(oldReaderName);
-            }
-        }
-
-
-        private void VerifyFabricClientCreateCalls(params string[] serviceNames)
-        {
-            _fabricClientMock.Verify(c => c.CreateServiceAsync(It.IsAny<ServiceCreationDescription>(), It.IsAny<CancellationToken>()), Times.Exactly(serviceNames.Length));
-
-            foreach (var serviceName in serviceNames)
-            {
-                _fabricClientMock.Verify(c => c.CreateServiceAsync(
-                        It.Is<ServiceCreationDescription>(m => m.ServiceName == serviceName),
-                        It.IsAny<CancellationToken>()),
-                    Times.Once);
-            }
-        }
-
-        private void VerifyFabricClientDeleteCalls(params string[] serviceNames)
-        {
-            _fabricClientMock.Verify(c => c.DeleteServiceAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Exactly(serviceNames.Length));
-
-            foreach (var serviceName in serviceNames)
-            {
-                _fabricClientMock.Verify(c => c.DeleteServiceAsync(It.Is<string>(m => m == serviceName), It.IsAny<CancellationToken>()), Times.Once);
-            }
-        }
-
-        private void VerifyServiceCreatedEventPublished(params string[] serviceNames)
-        {
-            _bigBrotherMock.Verify(b => b.Publish(It.IsAny<ReaderServiceCreatedEvent>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<int>()), Times.Exactly(serviceNames.Length));
-
-            foreach (var serviceName in serviceNames)
-            {
-                _bigBrotherMock.Verify(b => b.Publish(
-                        It.Is<ReaderServiceCreatedEvent>(m => m.ReaderName == serviceName),
-                        It.IsAny<string>(),
-                        It.IsAny<string>(),
-                        It.IsAny<int>()),
-                    Times.Once);
-            }
-        }
-
-        private void VerifyServiceDeletedEventPublished(params string[] serviceNames)
-        {
-            _bigBrotherMock.Verify(b => b.Publish(
-                   It.IsAny<ReaderServicesDeletionEvent>(),
-                   It.IsAny<string>(),
-                   It.IsAny<string>(),
-                   It.IsAny<int>()),
-                Times.Exactly(serviceNames.Length > 0 ? 1 : 0));
-
-            foreach (var serviceName in serviceNames)
-            {
-                _bigBrotherMock.Verify(b => b.Publish(
-                       It.Is<ReaderServicesDeletionEvent>(m => m.DeletedNames.Contains(serviceName) || m.Failed.Contains(serviceName)),
-                       It.IsAny<string>(),
-                       It.IsAny<string>(),
-                       It.IsAny<int>()),
-                    Times.Once);
             }
         }
     }
